test(user enrol request): add and fix more tests
