+++ app/controllers/instance_user_role_requests_controller.rb
@@ -12,7 +12,7 @@ class InstanceUserRoleRequestsController < ApplicationController
   def new
     @existing_role_request = @user_role_request.
                              instance.user_role_requests.
-                             where(creator_id: current_user.id, workflow_state: :pending).first
+                             where(user_id: current_user.id, workflow_state: :pending).first
     if @existing_role_request
       redirect_to edit_instance_user_role_request_path(@existing_role_request)
     else
