+++ spec/features/course/enrol_request_management_spec.rb
@@ -5,7 +5,7 @@ RSpec.feature 'Course: EnrolRequests' do
   let(:instance) { Instance.default }
 
   with_tenant(:instance) do
-    let(:course) { create(:course, default_timeline_algorithm: 'fomo') }
+    let(:course) { create(:course) }
     let(:user) { create(:course_manager, course: course).user }
     let!(:enrol_request) { create(:course_enrol_request, course: course) }
     before { login_as(user, scope: :user) }
@@ -18,9 +18,13 @@ RSpec.feature 'Course: EnrolRequests' do
       end
 
       wait_for_ajax
+      enrol_request.reload
+      expect(enrol_request.workflow_state).to eq('approved')
+
       course_user = course.course_users.find_by(user_id: enrol_request.user.id)
       expect(course_user).to be_present
-      expect(course_user.timeline_algorithm).to eq('fomo')
+
+      expect(current_path).to eq(course_enrol_requests_path(course))
       expect(page).to have_no_content_tag_for(enrol_request)
       expect(page).to have_text('course.enrol_requests.approve.success')
 
@@ -28,21 +32,23 @@ RSpec.feature 'Course: EnrolRequests' do
       expect(page).to have_content_tag_for(course_user)
     end
 
-    # Allow user to request to enrol again in case she neglects to enter her
-    # invitation code the first time
-    scenario 'Course staff can delete request' do
+    scenario 'Course staff can reject request', js: true do
       visit course_enrol_requests_path(course)
       expect(page).to have_field('course_user_name', with: enrol_request.user.name)
 
       expect do
         within find(content_tag_selector(enrol_request)) do
-          find_link(nil, href: course_enrol_request_path(course, enrol_request)).click
+          find_link(nil, href: reject_course_enrol_request_path(course, enrol_request)).click
         end
-      end.to change(course.enrol_requests, :count).by(-1)
+      end.to change(course.enrol_requests, :count).by(0)
 
+      enrol_request.reload
+
+      expect(enrol_request.workflow_state).to eq('rejected')
       expect(current_path).to eq(course_enrol_requests_path(course))
+      expect(page).to have_no_content_tag_for(enrol_request)
       expect(page).to have_selector('div.alert',
-                                    text: I18n.t('course.enrol_requests.destroy.success'))
+                                    text: I18n.t('course.enrol_requests.reject.success'))
     end
   end
 end
