+++ spec/features/instance_user_role_requests_management_spec.rb
@@ -40,6 +40,36 @@ RSpec.feature 'Instance::UserRoleRequests' do
 
         expect(request.reload.reason).to eq(new_reason)
       end
+
+      scenario 'when there is an existing user request, I am redirected to the edit page', js: true do
+        request = create(:role_request, user: user, instance: instance)
+        visit new_instance_user_role_request_path
+        expect(current_path).to eq(edit_instance_user_role_request_path(request))
+      end
+
+      scenario 'when I access a approved user request edit page', js: true do
+        request = create(:role_request, :approved, user: user, instance: instance)
+        visit edit_instance_user_role_request_path(request)
+        expect(page).to have_text(I18n.t('instance_user_role_requests.form.role_request_approved'))
+
+        expect(page).to have_field('user_role_request[organization]', visible: false, disabled: true)
+        expect(page).to have_field('user_role_request[designation]', visible: false, disabled: true)
+        expect(page).to have_field('user_role_request[reason]', visible: false, disabled: true)
+
+        expect(page).not_to have_button(I18n.t('instance_user_role_requests.form.edit_button'))
+      end
+
+      scenario 'when I access a rejected user request edit page', js: true do
+        request = create(:role_request, :rejected, user: user, instance: instance)
+        visit edit_instance_user_role_request_path(request)
+        expect(page).to have_text(I18n.t('instance_user_role_requests.form.role_request_rejected'))
+
+        expect(page).to have_field('user_role_request[organization]', visible: false, disabled: true)
+        expect(page).to have_field('user_role_request[designation]', visible: false, disabled: true)
+        expect(page).to have_field('user_role_request[reason]', visible: false, disabled: true)
+
+        expect(page).not_to have_button(I18n.t('instance_user_role_requests.form.edit_button'))
+      end
     end
 
     context 'As an instance admin' do
